feat: move token-related exceptions to auth package for better organization
