+++ backend/src/main/java/com/cinerama/backend/exception/auth/ExpiredTokenException.java
@@ -1,4 +1,4 @@
-package com.cinerama.backend.exception;
+package com.cinerama.backend.exception.auth;
 
 public class ExpiredTokenException extends RuntimeException {
     public ExpiredTokenException(String message) {
