+++ backend/src/main/java/com/cinerama/backend/exception/auth/TokenNotBelongUserException.java
@@ -1,4 +1,4 @@
-package com.cinerama.backend.exception;
+package com.cinerama.backend.exception.auth;
 
 public class TokenNotBelongUserException extends RuntimeException {
     public TokenNotBelongUserException(String message) {
